Port complex ordering test to dag-map.
