+++ packages/babel-core/test/api.js
@@ -2,7 +2,6 @@ import * as babel from "../lib/index";
 import buildExternalHelpers from "../lib/tools/build-external-helpers";
 import sourceMap from "source-map";
 import assert from "assert";
-import path from "path";
 import Plugin from "../lib/config/plugin";
 import generator from "babel-generator";
 
@@ -191,158 +190,6 @@ describe("api", function () {
 
     assert.equal(calledRaw, 4);
     assert.equal(calledIntercept, 4);
-  });
-
-  it("pass per preset", function () {
-    let aliasBaseType = null;
-
-    function execTest(passPerPreset) {
-      return babel.transform("type Foo = number; let x = (y): Foo => y;", {
-        passPerPreset: passPerPreset,
-        presets: [
-          // First preset with our plugin, "before"
-          function () {
-            return {
-              plugins: [
-                new Plugin({
-                  visitor: {
-                    Function: function (path) {
-                      const alias = path.scope.getProgramParent().path.get("body")[0].node;
-                      if (!babel.types.isTypeAlias(alias)) return;
-
-                      // In case of `passPerPreset` being `false`, the
-                      // alias node is already removed by Flow plugin.
-                      if (!alias) {
-                        return;
-                      }
-
-                      // In case of `passPerPreset` being `true`, the
-                      // alias node should still exist.
-                      aliasBaseType = alias.right.type; // NumberTypeAnnotation
-                    },
-                  },
-                }),
-              ],
-            };
-          },
-
-          // ES2015 preset
-          require(__dirname + "/../../babel-preset-es2015"),
-
-          // Third preset for Flow.
-          function () {
-            return {
-              plugins: [
-                require(__dirname + "/../../babel-plugin-syntax-flow"),
-                require(__dirname + "/../../babel-plugin-transform-flow-strip-types"),
-              ],
-            };
-          },
-        ],
-      });
-    }
-
-    // 1. passPerPreset: true
-
-    let result = execTest(true);
-
-    assert.equal(aliasBaseType, "NumberTypeAnnotation");
-
-    assert.deepEqual([
-      "\"use strict\";",
-      "",
-      "var x = function x(y) {",
-      "  return y;",
-      "};",
-    ].join("\n"), result.code);
-
-    // 2. passPerPreset: false
-
-    aliasBaseType = null;
-
-    result = execTest(false);
-
-    assert.equal(aliasBaseType, null);
-
-    assert.deepEqual([
-      "\"use strict\";",
-      "",
-      "var x = function x(y) {",
-      "  return y;",
-      "};",
-    ].join("\n"), result.code);
-
-  });
-
-  it("complex plugin and preset ordering", function() {
-    function pushPlugin(str) {
-      return {
-        visitor: {
-          Program(path) {
-            path.pushContainer("body", babel.types.expressionStatement(babel.types.identifier(str)));
-          },
-        },
-      };
-    }
-
-    function pushPreset(str) {
-      return { plugins: [pushPlugin(str)] };
-    }
-
-    const result = babel.transform("", {
-      filename: path.join(__dirname, "fixtures", "config", "complex-plugin-config", "file.js"),
-      presets: [
-        pushPreset("argone"),
-        pushPreset("argtwo"),
-      ],
-      env: {
-        development: {
-          passPerPreset: true,
-          presets: [
-            pushPreset("argthree"),
-            pushPreset("argfour"),
-          ],
-          env: {
-            development: {
-              passPerPreset: true,
-              presets: [
-                pushPreset("argfive"),
-                pushPreset("argsix"),
-              ],
-            },
-          },
-        },
-      },
-    });
-
-    assert.equal(result.code, [
-      "argtwo;",
-      "argone;",
-      "eleven;",
-      "twelve;",
-      "one;",
-      "two;",
-      "five;",
-      "six;",
-      "three;",
-      "four;",
-      "seventeen;",
-      "eighteen;",
-      "nineteen;",
-      "twenty;",
-      "thirteen;",
-      "fourteen;",
-      "fifteen;",
-      "sixteen;",
-      "argfive;",
-      "argsix;",
-      "argthree;",
-      "argfour;",
-      "seven;",
-      "eight;",
-      "nine;",
-      "ten;",
-    ].join("\n"));
   });
 
   it("source map merging", function () {
