+++ packages/babel-core/test/fixtures/config/complex-plugin-config/eighteen.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("eighteen");
+module.exports = require("./plugin")("eighteen", ["eighteen"], ["nineteen"]);
