+++ packages/babel-core/test/fixtures/config/complex-plugin-config/fifteen.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("fifteen");
+module.exports = require("./plugin")("fifteen", ["fifteen"], ["sixteen"]);
