+++ packages/babel-core/test/fixtures/config/complex-plugin-config/five.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("five");
+module.exports = require("./plugin")("five", ["five"], ["six"]);
