+++ packages/babel-core/test/fixtures/config/complex-plugin-config/four.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("four");
+module.exports = require("./plugin")("four", ["four"], ["five"]);
