+++ packages/babel-core/test/fixtures/config/complex-plugin-config/nine.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("nine");
+module.exports = require("./plugin")("nine", ["nine"], ["ten"]);
