+++ packages/babel-core/test/fixtures/config/complex-plugin-config/nineteen.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("nineteen");
+module.exports = require("./plugin")("nineteen", ["nineteen"], ["twenty"]);
