+++ packages/babel-core/test/fixtures/config/complex-plugin-config/plugin.js
@@ -1,10 +1,13 @@
-module.exports = function(str) {
+module.exports = function(name, capabilities, dependencies) {
   return function(babel) {
     return {
+      name: name,
+      capabilities: capabilities,
+      dependencies: dependencies,
       visitor: {
         Program: function(path) {
           path.pushContainer("body", [
-            babel.types.expressionStatement(babel.types.identifier(str)),
+            babel.types.expressionStatement(babel.types.identifier(name)),
           ]);
         },
       },
