+++ packages/babel-core/test/fixtures/config/complex-plugin-config/seven.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("seven");
+module.exports = require("./plugin")("seven", ["seven"], ["eight"]);
