+++ packages/babel-core/test/fixtures/config/complex-plugin-config/six.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("six");
+module.exports = require("./plugin")("six", ["six"], ["seven"]);
