+++ packages/babel-core/test/fixtures/config/complex-plugin-config/sixteen.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("sixteen");
+module.exports = require("./plugin")("sixteen", ["sixteen"], ["seventeen"]);
