+++ packages/babel-core/test/fixtures/config/complex-plugin-config/twenty.js
@@ -1,1 +1,1 @@
-module.exports = require("./plugin")("twenty");
+module.exports = require("./plugin")("twenty", ["twenty"], []);
